Move parse cell data into separate file
